requirements.txt
click>=8.0.0
PyYAML>=5.4.1
requests>=2.25.1
ratelimiter>=1.2.0
azure-devops>=6.0.0b3
msrest>=0.6.21
